feat: move analytics map overlay component to analytics layout for better colocation
